+++ src/shapes/Image.ts
@@ -9,25 +9,49 @@
  * Description:
  */
 
-import { Util } from '../Util';
-import { Factory } from '../Factory';
+import { Util }               from '../Util';
+import { Factory }            from '../Factory';
 import { Shape, ShapeConfig } from '../Shape';
 import { getNumberValidator } from '../Validators';
-import { _registerNode } from '../Global';
+import { _registerNode }      from '../Global';
 
-import { GetSet, IRect } from '../types';
-import { Context } from '../Context';
+import { GetSet, IRect }  from '../types';
+import { SceneContext }   from '../Context';
+import { Size2D, sizeOf } from '../common/Size2D';
+
+export enum ImageSourcePolicy {
+  DATAURL,
+  URL,
+}
 
 export interface ImageConfig extends ShapeConfig {
-  image: CanvasImageSource | undefined;
+  /**
+   * Image data to draw
+   */
+  image?: CanvasImageSource | undefined;
+
+  /**
+   * Cropping rect to apply
+   */
   crop?: IRect;
+
+  /**
+   * Source data to use (can be url or data url)
+   * Useful for persistence purposes
+   */
+  sourceData?: string;
+
+  /**
+   * Type of source data content
+   */
+  sourcePolicy?: ImageSourcePolicy;
 }
 
 /**
  * Image constructor
  * @constructor
- * @memberof Konva
- * @augments Konva.Shape
+ * @memberof Pamela
+ * @augments Pamela.Shape
  * @param {Object} config
  * @param {Image} config.image
  * @param {Object} [config.crop]
@@ -36,7 +60,7 @@ export interface ImageConfig extends ShapeConfig {
  * @example
  * var imageObj = new Image();
  * imageObj.onload = function() {
- *   var image = new Konva.Image({
+ *   var image = new Pamela.Image({
  *     x: 200,
  *     y: 50,
  *     image: imageObj,
@@ -55,6 +79,7 @@ export class Image extends Shape<ImageConfig> {
 
     this._setImageLoad();
   }
+
   _setImageLoad() {
     const image = this.image() as any;
     // check is image is already loaded
@@ -71,15 +96,26 @@ export class Image extends Shape<ImageConfig> {
       });
     }
   }
+
   _useBufferCanvas() {
     return super._useBufferCanvas(true);
   }
-  _sceneFunc(context: Context) {
+
+  _sceneFunc(context: SceneContext) {
     const width = this.getWidth();
     const height = this.getHeight();
     const image = this.attrs.image;
     let params;
 
+    // Load data url if any
+    if (!image && this.sourcePolicy() === ImageSourcePolicy.DATAURL && this.sourceData())
+      this.loadFromDataURL(this.sourceData());
+
+    // Load from url if any
+    if (!image && this.sourcePolicy() === ImageSourcePolicy.URL && this.sourceData())
+      this.loadFromURL(this.sourceData());
+
+    // Draw image
     if (image) {
       const cropWidth = this.attrs.cropWidth;
       const cropHeight = this.attrs.cropHeight;
@@ -113,6 +149,52 @@ export class Image extends Shape<ImageConfig> {
   }
 
   /**
+   * Loads an image from the url. Keeps all the other attributes and changes only
+   * the content.
+   * @param url Url to load from
+   * @param cors Cross origin headers to apply to request
+   */
+  public loadFromURL(url: string, cors?: string) {
+    var img = Util.createImageElement();
+    const th = this;
+    img.onload = () => {
+      th.attrs.image = img;
+      this._requestDraw();
+    };
+    img.src = url;
+  }
+
+  /**
+   * Loads an image from data url
+   * @param dataUrl Image data url
+   */
+  public loadFromDataURL(dataUrl: string) {
+    var img = Util.createImageHelper();
+    let th = this;
+    img.onload = function () {
+      th.attrs.image = img;
+      th._requestDraw();
+    };
+    img.src = dataUrl;
+  }
+
+  /**
+   * Loads an image starting from the data url
+   * @param dataURL
+   * @param callback
+   */
+  public static fromDataURL(dataURL: string, callback: (img: Image) => void) {
+    var img = Util.createImageHelper();
+    let th = this;
+    img.onload = function () {
+      callback(new Image({
+        image: img,
+      }));
+    };
+    img.src = dataURL;
+  }
+
+  /**
    * Function called when we have to detect hit boundaries.
    * It should draw a rectangle with specific size of this
    * shape using its fill color.
@@ -127,37 +209,79 @@ export class Image extends Shape<ImageConfig> {
     context.closePath();
     context.fillStrokeShape(this);
   }
+
+  /**
+   * Returns the original size of this image, without crop and transform
+   */
+  public getOriginalSize(): Size2D | undefined {
+    if (this.attrs.image) {
+      return sizeOf(this.attrs.image.width, this.attrs.image.height);
+    } else return undefined;
+  }
+
+  /**
+   * Returns the original width
+   */
+  public getOriginalWidth(): number | undefined {
+    return this.getOriginalSize()?.getWidth();
+  }
+
+  /**
+   * Returns the original height of this image
+   */
+  public getOriginalHeight(): number | undefined {
+    return this.getOriginalSize()?.getHeight();
+  }
+
+  /**
+   * Returns actual shape width. To get original image width use getOriginalSize().getWidth()
+   * method
+   */
   getWidth() {
     return this.attrs.width ?? this.image()?.width;
   }
+
+  /**
+   * Returns actual shape height.
+   * To get original image width use getOriginalSize().getHeight()
+   */
   getHeight() {
     return this.attrs.height ?? this.image()?.height;
   }
 
   /**
-   * load image from given url and create `Konva.Image` instance
+   * load image from given url and create `Pamela.Image` instance
    * @method
-   * @memberof Konva.Image
+   * @memberof Pamela.Image
    * @param {String} url image source
-   * @param {Function} callback with Konva.Image instance as first argument
+   * @param {Function} callback with Pamela.Image instance as first argument
+   * @param {String} cors Cross origin permissions to download this image. By default it is `Anonymous`
    * @example
-   *  Konva.Image.fromURL(imageURL, function(image){
-   *    // image is Konva.Image instance
+   *  Pamela.Image.fromURL(imageURL, image => {
+   *    // image is Pamela.Image instance
    *    layer.add(image);
    *    layer.draw();
    *  });
    */
-  static fromURL(url, callback) {
+  static fromURL(url: string, callback: (image: Image) => void, cors?: string) {
     var img = Util.createImageElement();
     img.onload = function () {
       var image = new Image({
         image: img,
+        src: url,
       });
       callback(image);
     };
-    img.crossOrigin = 'Anonymous';
+    img.crossOrigin = cors;
     img.src = url;
   }
+
+  // toDataURL(config?: { x?: number; y?: number; width?: number; height?: number; pixelRatio?: number; mimeType?: string; quality?: number; callback?: (str: string) => void }): string {
+  //   console.log("Reac");
+  //   if(this.attrs.image) {
+  //     return "";
+  //   }
+  // }
 
   image: GetSet<CanvasImageSource | undefined, this>;
   crop: GetSet<IRect, this>;
@@ -165,13 +289,15 @@ export class Image extends Shape<ImageConfig> {
   cropY: GetSet<number, this>;
   cropWidth: GetSet<number, this>;
   cropHeight: GetSet<number, this>;
+  sourceData: GetSet<string, this>;
+  sourcePolicy: GetSet<ImageSourcePolicy, this>;
 }
 
 Image.prototype.className = 'Image';
 _registerNode(Image);
 /**
  * get/set image source. It can be image, canvas or video element
- * @name Konva.Image#image
+ * @name Pamela.Image#image
  * @method
  * @param {Object} image source
  * @returns {Object}
@@ -188,7 +314,7 @@ Factory.addComponentsGetterSetter(Image, 'crop', ['x', 'y', 'width', 'height']);
 /**
  * get/set crop
  * @method
- * @name Konva.Image#crop
+ * @name Pamela.Image#crop
  * @param {Object} crop
  * @param {Number} crop.x
  * @param {Number} crop.y
@@ -212,7 +338,7 @@ Factory.addGetterSetter(Image, 'cropX', 0, getNumberValidator());
 /**
  * get/set crop x
  * @method
- * @name Konva.Image#cropX
+ * @name Pamela.Image#cropX
  * @param {Number} x
  * @returns {Number}
  * @example
@@ -226,7 +352,7 @@ Factory.addGetterSetter(Image, 'cropX', 0, getNumberValidator());
 Factory.addGetterSetter(Image, 'cropY', 0, getNumberValidator());
 /**
  * get/set crop y
- * @name Konva.Image#cropY
+ * @name Pamela.Image#cropY
  * @method
  * @param {Number} y
  * @returns {Number}
@@ -241,7 +367,7 @@ Factory.addGetterSetter(Image, 'cropY', 0, getNumberValidator());
 Factory.addGetterSetter(Image, 'cropWidth', 0, getNumberValidator());
 /**
  * get/set crop width
- * @name Konva.Image#cropWidth
+ * @name Pamela.Image#cropWidth
  * @method
  * @param {Number} width
  * @returns {Number}
@@ -254,9 +380,10 @@ Factory.addGetterSetter(Image, 'cropWidth', 0, getNumberValidator());
  */
 
 Factory.addGetterSetter(Image, 'cropHeight', 0, getNumberValidator());
+
 /**
  * get/set crop height
- * @name Konva.Image#cropHeight
+ * @name Pamela.Image#cropHeight
  * @method
  * @param {Number} height
  * @returns {Number}
@@ -267,3 +394,7 @@ Factory.addGetterSetter(Image, 'cropHeight', 0, getNumberValidator());
  * // set crop height
  * image.cropHeight(20);
  */
+
+Factory.addGetterSetter(Image, 'sourcePolicy');
+
+Factory.addGetterSetter(Image, 'sourceData');